move pseudolegal check to move impl
